chore: move randomTable util back to fixtures
